pyproject.toml
[project]
name = "backgrounder"
version = "0.1.0"
description = "AI-powered background check agent — LinkedIn, GitHub, social media, resume verification, reverse photo search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.48.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.7.0",
]

[project.scripts]
backgrounder = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py311"
